+++ amr_runtime_observation/src/main.cpp
@@ -1,0 +1,12 @@
+#include "amr_runtime_observation/runtime_observation.hpp"
+
+int main(int argc, char ** argv)
+{
+  rclcpp::init(argc, argv);
+  auto node = std::make_shared<amr::runtime::observation::RuntimeObservation>();
+  rclcpp::executors::SingleThreadedExecutor executor;
+  executor.add_node(node);
+  executor.spin();
+  rclcpp::shutdown();
+  return 0;
+}